Backend/requirements.txt
fastapi
uvicorn
torch
segmentation-models-pytorch
Pillow
requests
numpy